Dividing code into components and build basic layout structure
