+++ src/pages/Dashboard.js
@@ -1,0 +1,11 @@
+export default function Dashboard() {
+  return (
+    <>
+      <h1>使用者名稱</h1>
+      <div>
+        <h2>筆記們的統計資料</h2>
+        <p>資料內容</p>
+      </div>
+    </>
+  );
+}
