+++ src/pages/TryOut.js
@@ -1,0 +1,25 @@
+/* eslint-disable react/button-has-type */
+/* eslint-disable react/prop-types */
+
+// export default function TryOut() {
+//   function handleStoreItem() {
+//     localStorage.setItem('myData', JSON.stringify(cards));
+//   }
+//   return (
+//     <>
+//       <button
+//         onClick={handleStoreItem}
+//       >
+//         儲存到 localStorage
+//       </button>
+//       <button
+//         onClick={() => {
+//           const newData = JSON.parse(localStorage.getItem('myData'));
+//           setCards(newData);
+//         }}
+//       >
+//         從 localStorage 載入我的筆記
+//       </button>
+//     </>
+//   );
+// }
